+++ components/Quiz.vue
@@ -1,7 +1,17 @@
 <script setup>
-import QuestionsSteps from "./QuestionsSteps.vue"
-import ResultBox from "./ResultBox.vue"
+import QuestionsSteps from "./quiz/QuestionsSteps.vue"
+import ResultBox from "./quiz/ResultBox.vue"
 import { ref, computed } from "vue"
+import quiz1 from "./quiz1.json"
+
+const props = defineProps({
+    questionsNumber: {
+        type: Number,
+        required: true
+    }
+})
+
+const questions = [quiz1]
 
 const currentStep = ref('questioning')
 const currentComponent = computed(() => ({
@@ -14,6 +24,9 @@ const result = ref({})
 const bindData = computed(() => ({
     'result': {
         result: result.value
+    },
+    'questioning': {
+        questions: questions[props.questionsNumber - 1]
     }
 }[currentStep.value]))
