+++ components/quiz/ResultBox.vue
