+++ components/quiz/utils/index.js
@@ -1,7 +1,3 @@
-export function generateRndColor() {
-    return '#' + Math.floor(Math.random() * 16777215).toString(16)
-}
-
 export function shuffleArray(arr) {
     const array = [...arr]
     for (let i = array.length - 1; i > 0; i--) {
@@ -13,8 +9,4 @@ export function shuffleArray(arr) {
 
 export function asyncTimeout(ms) {
     return new Promise(resolve => setTimeout(() => resolve(), ms))
-}
-
-export function subtractDates(date1, date2) {
-    return (date2.getTime() - date1.getTime()) / 1000
 }
