+++ components/quiz1.json
@@ -1,0 +1,37 @@
+[
+  {
+    "type": "multiple",
+    "difficulty": "hard",
+    "question": "Question Statement",
+    "correct_answer": "Correct1",
+    "incorrect_answers": ["Wrong1", "Wrong2", "Wrong3"]
+  },
+  {
+    "type": "multiple",
+    "difficulty": "easy",
+    "question": "Question Statement",
+    "correct_answer": "Correct2",
+    "incorrect_answers": ["Two words", "And maybe more", "Wrong6"]
+  },
+  {
+    "type": "multiple",
+    "difficulty": "easy",
+    "question": "Question Statement",
+    "correct_answer": "Correct3",
+    "incorrect_answers": ["Wrong7", "Wrong8", "Wrong9"]
+  },
+  {
+    "type": "multiple",
+    "difficulty": "easy",
+    "question": "Question Statement",
+    "correct_answer": "Correct4",
+    "incorrect_answers": ["Wrong10", "Wrong11", "Wrong12"]
+  },
+  {
+    "type": "boolean",
+    "difficulty": "easy",
+    "question": "Question Statement",
+    "correct_answer": "True",
+    "incorrect_answers": ["False"]
+  }
+]
