+++ pages/01-introducao/introducao-2.md
@@ -1,6 +1,6 @@
 # Quiz
 
-<Quiz />
+<Quiz questionsNumber="1" />
 
 ---
 layout: two-cols-center
